# perf: batch validation of paths