--- a/lib/validator.js
+++ b/lib/validator.js
@@ -9,7 +9,8 @@ function validator (opts = {}) {
   } = opts
 
   return function validate ({ paths }) {
-    paths.forEach((s) => {
+    const proxy = new Proxy({}, { get: () => proxy, set: () => { throw Error() } })
+    const exprsToCheck = paths.map((s, i) => {
       if (typeof s !== 'string') {
         throw Error(ERR_PATHS_MUST_BE_STRINGS())
       }
@@ -18,16 +19,33 @@ function validator (opts = {}) {
         const expr = (s[0] === '[' ? '' : '.') + s.replace(/^\*/, '〇').replace(/\.\*/g, '.〇').replace(/\[\*\]/g, '[〇]')
         if (/\n|\r|;/.test(expr)) throw Error()
         if (/\/\*/.test(expr)) throw Error()
-        /* eslint-disable-next-line */
-        Function(`
-            'use strict'
-            const o = new Proxy({}, { get: () => o, set: () => { throw Error() } });
-            const 〇 = null;
-            o${expr}
-            if ([o${expr}].length !== 1) throw Error()`)()
+
+        return `
+        context.〇〇 = \`${s.replace(/`/g, '\\`')}\`
+        o${expr}
+        if ([o${expr}].length !== 1) return;
+        `
       } catch (e) {
         throw Error(ERR_INVALID_PATH(s))
       }
     })
+
+    const o = new Proxy({}, { get: () => o, set: () => { throw Error() } })
+    const context = { 〇〇: paths[0] }
+
+    try {
+      /* eslint-disable-next-line */
+      Function('〇', 'o', 'context', `
+        'use strict'
+
+        ${exprsToCheck.join('\n')}
+
+        context.〇〇 = null;
+      `)(null, o, context)
+
+      if (context.〇〇 !== null) { throw Error(ERR_INVALID_PATH(context.〇〇)) }
+    } catch (e) {
+      throw Error(ERR_INVALID_PATH(context.〇〇))
+    }
   }
 }