Hello o/

I saw the construction of the `fastRedact` is a little bit slow, in my use case, it took 136ms.

Digging a little bit and I found the main problem was with `createContext` that is used inside `validator.js`.

For each path, it calls `createContext` (the most expensive) and then `runInContext`.

To optimize, instead call it once per path, I join all calls in one script and try to avoid throwing errors inside the validation, instead, I just assign the current path that will be validated to a special variable and then read it in the final.

About the optimizations, below the old redact version:

```
Initializing and redacting x 398 ops/sec ±2.01% (85 runs sampled)
Redacting with cached x 2,095,357 ops/sec ±1.39% (88 runs sampled)
```

And with this new way:

```
Initializing and redacting x 1,257 ops/sec ±3.75% (75 runs sampled)
Redacting with cached x 2,103,045 ops/sec ±1.31% (87 runs sampled)
```

<details>
<summary>benchmark.js</summary>

```js
const Benchmark = require('benchmark')
const suite = new Benchmark.Suite()

const options = {
  paths: ['*.authorization', '*.*.authorization', '*.cookie', '*.*.cookie']
};
const cachedRedact = require('./index')(options)

suite
  .add('Initializing and redacting', function () {
    const fastRedact = require('./index')

    fastRedact(options)({ a: 'test' })
  })
  .add('Redacting with cached', function () {
    cachedRedact({ a: 'test' })
  })
  .on('cycle', function (event) {
    console.log(String(event.target))
  })
  .run({ 'async': false })
```

</details>

As you can see above, the new way initializes ~3x faster than the old version.