@H4ad 2023-03-23
I see the tests failing for nodejs 6.x but here's an approximate example of the output of the script that will run on the VM:

```js
var o = new Proxy({}, { get: () => o, set: () => { throw Error() } });

(function () {
  'use strict';

  〇〇 = '*.authorization';
  o.〇.authorization
  if ([o.〇.authorization].length !== 1) return;
  

  〇〇 = '*.*.authorization';
  o.〇.〇.authorization
  if ([o.〇.〇.authorization].length !== 1) return;
  

  〇〇 = '*.cookie';
  o.〇.cookie
  if ([o.〇.cookie].length !== 1) return;
  

  〇〇 = '*.*.cookie';
  o.〇.〇.cookie
  if ([o.〇.〇.cookie].length !== 1) return;
  


  〇〇 = null;
}).bind({
  o,
  〇: null,
  〇〇: '*.authorization'
})()
```

Running on NodeJS 6.x throws the error: "ReferenceError: 〇〇 is not defined", but I don't have any idea why.

Do you guys have some hints about this issue?