@mcollina 2023-03-23
I think we can cut support for a lot od those old node majors.

wdyt @davidmarkclements?