@jsumners 2023-03-23
I would drop support for all of the versions currently being tested in CI. There isn't enough life left in 14 to justify the maintenance burden.