@H4ad 2023-05-15
After merging #53, the performance was greatly improved by the remove of `vm`, now, I don't think this PR will bring too much perf improvement, here's the new benchmark:

```
Initializing and redacting with oldRedact x 11,933 ops/sec ±1.35% (81 runs sampled)
Redacting with cached oldRedact x 1,002,374 ops/sec ±0.54% (92 runs sampled)

Initializing and redacting with newRedact x 12,454 ops/sec ±0.92% (90 runs sampled)
Redacting with cached newRedact x 918,342 ops/sec ±0.53% (93 runs sampled)
```
> The result of redacting the cache instance is very unstable for unknown reason.